pyproject.toml
[tool.poetry]
name = "wattstor-meteo"
version = "0.1.0"
description = ""
authors = ["silox <edruisl99@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
dependency-injector = "^4.43.0"
open-meteo = "^0.3.1"
fastapi-utils = "^0.8.0"
typing-inspect = "^0.9.0"
influxdb-client = {extras = ["async"], version = "^1.47.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
